+++ tests/lbm_mesapd_coupling/partially_saturated_cells_method/SettlingSpherePSM.cpp
@@ -366,22 +366,22 @@ int main(int argc, char** argv)
       }
       if (std::strcmp(argv[i], "--vtkIOFreq") == 0)
       {
-         vtkIOFreq = uint_c(std::atof(argv[++i]));
+         vtkIOFreq = uint_c(std::stoul(argv[++i]));
          continue;
       }
       if (std::strcmp(argv[i], "--fluidType") == 0)
       {
-         fluidType = uint_c(std::atof(argv[++i]));
+         fluidType = uint_c(std::stoul(argv[++i]));
          continue;
       }
       if (std::strcmp(argv[i], "--numRPDSubCycles") == 0)
       {
-         numRPDSubCycles = uint_c(std::atof(argv[++i]));
+         numRPDSubCycles = uint_c(std::stoul(argv[++i]));
          continue;
       }
       if (std::strcmp(argv[i], "--resolution") == 0)
       {
-         numberOfCellsInHorizontalDirection = uint_c(std::atof(argv[++i]));
+         numberOfCellsInHorizontalDirection = uint_c(std::stoul(argv[++i]));
          continue;
       }
       if (std::strcmp(argv[i], "--noForceAveraging") == 0)
